examples/github/requirements.txt
fastmcp
python-dotenv
google-genai
streamlit
agent-dna